requirements.txt
ultralytics>=8.0.196
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0.76
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
loguru>=0.7.0
numpy>=1.26.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
pytest>=7.4.2
httpx>=0.24.1
PyYAML>=6.0.1
albumentations>=1.3.1
matplotlib>=3.7.2
types-PyYAML>=6.0.12.11